// components/calculators/math/trigonometry/trigonometry.types.ts
export type AngleUnit = "degree" | "radian"

export interface TrigResult {
  sin: number | string
  cos: number | string
  tan: number | string
  cot: number | string
  arcsin: number | string
  arccos: number | string
  arctan: number | string
  arccot: number | string
  conversion: number | string
}